requirements.txt
streamlit
requests
azure-storage-blob
azure-storage-queue
python-dotenv
